- added three buttons ingame - save, return to menu and exit - added functionality for return to menu and exit buttons - integrated buttons into ingame UI - created skeleton of save game function, more to be added in time - split Game class into multiple files to ease navigation - various small changes

+++ Settings.cs
@@ -23,8 +23,8 @@ namespace VNet
 		/*
 		 * Global variables which change at runtime
 		 */
-		public static int windowWidth = 1280;
-		public static int windowHeight = 720;
+		public static double windowWidth = 1280;
+		public static double windowHeight = 720;
 		public static bool textDisplayedFully;
 		public static bool executeNext;
 		public static bool inGame;
